Fix some weird issues with server.
It seems that when a node tried to connect to the same other node
at the same time on two different interfaces, the connections
would just lock up. Not quite sure why this happened, but I added
code to prevent the trigger condition from manifesting, which
seems to solve the problem.